requirements.txt
PyQt5
opencc-python-reimplemented
PyInstaller
pyinstaller-hooks-contrib
